Former-commit-id: e2a7d6b6cf89e53e4a3bcc4ad760791983f21024

+++ Projects/UnitTests/Classes/RectSpriteTest.cpp
@@ -175,7 +175,8 @@ void RectSpriteTest::TestFunction(TestTemplate<RectSpriteTest>::PerfFuncData *da
         if (images[i] && sprites[i])
         {
             Vector2 spriteSize = sprites[i]->GetSize();
-            Vector2 imageSize = Vector2(images[i]->GetWidth(), images[i]->GetHeight());
+            Vector2 imageSize = Vector2(static_cast<float32>(images[i]->GetWidth()), 
+				static_cast<float32>(images[i]->GetHeight()));
 
             if (spriteSize != imageSize)
             {
